Moving IC canister client in metrics mod
